Move 'v4orV5' util into 'clusterUtils'
